+++ inst/resources/init-rprofile.R
@@ -1,3 +1,3 @@
-#### -- Packrat Autoloader (version 0.4.1.10) -- ####
+#### -- Packrat Autoloader (version 0.4.1.11) -- ####
 source("packrat/init.R")
 #### -- End Packrat Autoloader -- ####